refactor: Separate the Model module into 2 submodules
Take Stat struct and put it in a different submodule than Model
