+++ src/lib.rs
@@ -1,5 +1,3 @@
-// Lib modules
-
 pub mod fight;
 pub mod global_test;
 mod math_tools;
@@ -7,7 +5,4 @@ pub mod model;
 pub mod modifier;
 pub mod prediction;
 pub mod regiment;
-
-// Binary modules
-
-pub mod web_server;
+pub mod stat;
